ui: Extract the timespan selector
